Make DG optional and add CMake support
Adds CMake support for finding Eigen, thus getting rid of absolute
paths. If Eigen is not in a standard path, you can use one of the
following CMake flags:
-DCMAKE_PREFIX_PATH=path/to/mypackages
-DEigen3_DIR=path/to/mypackages/share/eigen3/cmake/

Support for DG is enabled by default but can be disabled manually by
setting the CMake flag -DENABLE_DG=OFF. If Eigen is not found, support
is disabled automatically (regardless of the value of ENABLE_DG flag).